First attempt and re-writing a portion of the server in typescript.
