+++ packages/server/.eslintrc
@@ -3,6 +3,7 @@
     "emit": true,
     "key": true
   },
+  "parser": "@typescript-eslint/parser",
   "env": {
     "node": true
   },
