+++ packages/server/Dockerfile
@@ -9,6 +9,7 @@ ENV BUDIBASE_ENVIRONMENT=PRODUCTION
 # copy files and install dependencies
 COPY . ./
 RUN yarn
+RUN yarn build
 
 EXPOSE 4001
