+++ packages/server/package.json
@@ -3,22 +3,25 @@
   "email": "hi@budibase.com",
   "version": "0.9.63",
   "description": "Budibase Web Server",
-  "main": "src/electron.js",
+  "main": "src/index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/Budibase/budibase.git"
   },
   "scripts": {
+    "build": "tsc",
     "test": "jest --testPathIgnorePatterns=routes && yarn run test:integration",
     "test:integration": "jest --coverage --detectOpenHandles",
     "test:watch": "jest --watch",
     "build:docker": "docker build . -t app-service",
-    "run:docker": "node src/index",
+    "run:docker": "node dist/src/index.js",
     "dev:stack:up": "node scripts/dev/manage.js up",
     "dev:stack:down": "node scripts/dev/manage.js down",
     "dev:stack:nuke": "node scripts/dev/manage.js nuke",
-    "dev:builder": "yarn run dev:stack:up && nodemon src/index.js",
+    "dev:builder": "yarn run dev:stack:up && ts-node src/index.ts",
+    "format": "prettier --config ../../.prettierrc.json 'src/**/*.ts' --write",
     "lint": "eslint --fix src/",
+    "lint:fix": "yarn run format && yarn run lint",
     "initialise": "node scripts/initialise.js"
   },
   "jest": {
@@ -112,6 +115,11 @@
     "@babel/preset-env": "^7.14.4",
     "@budibase/standard-components": "^0.9.63",
     "@jest/test-sequencer": "^24.8.0",
+    "@types/bull": "^3.15.1",
+    "@types/koa": "^2.13.3",
+    "@types/koa-router": "^7.4.2",
+    "@types/node": "^15.12.4",
+    "@typescript-eslint/parser": "^4.28.0",
     "babel-jest": "^27.0.2",
     "docker-compose": "^0.23.6",
     "eslint": "^6.8.0",
@@ -119,7 +127,10 @@
     "jest": "^24.8.0",
     "nodemon": "^2.0.4",
     "pouchdb-adapter-memory": "^7.2.1",
-    "supertest": "^4.0.2"
+    "prettier": "^2.3.1",
+    "supertest": "^4.0.2",
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.4"
   },
   "gitHead": "d1836a898cab3f8ab80ee6d8f42be1a9eed7dcdc"
 }
