--- packages/server/src/integrations/Integration.js
@@ -1,15 +1,0 @@
-exports.QUERY_TYPES = {
-  SQL: "sql",
-  JSON: "json",
-  FIELDS: "fields",
-}
-
-exports.FIELD_TYPES = {
-  STRING: "string",
-  BOOLEAN: "boolean",
-  NUMBER: "number",
-  PASSWORD: "password",
-  LIST: "list",
-  OBJECT: "object",
-  JSON: "json",
-}
