--- packages/server/src/integrations/base/constants.js
@@ -1,11 +1,0 @@
-exports.Operation = {
-  CREATE: "CREATE",
-  READ: "READ",
-  UPDATE: "UPDATE",
-  DELETE: "DELETE",
-}
-
-exports.SortDirection = {
-  ASCENDING: "ASCENDING",
-  DESCENDING: "DESCENDING",
-}
