+++ packages/server/src/integrations/base/sql.ts
@@ -1,9 +1,10 @@
-const { DataSourceOperation, SortDirection } = require("../../constants")
-
+import { Knex, knex } from "knex"
 const BASE_LIMIT = 5000
+import { QueryJson, SearchFilters, QueryOptions, SortDirection, Operation } from "./definitions"
 
-function addFilters(query, filters) {
-  function iterate(structure, fn) {
+
+function addFilters(query: any, filters: SearchFilters | undefined): Knex.QueryBuilder {
+  function iterate(structure: { [key: string]: any }, fn: (key: string, value: any) => void) {
     for (let [key, value] of Object.entries(structure)) {
       fn(key, value)
     }
@@ -12,7 +13,7 @@ function addFilters(query, filters) {
     return query
   }
   // if all or specified in filters, then everything is an or
-  const allOr = !!filters.allOr
+  const allOr = filters.allOr
   if (filters.string) {
     iterate(filters.string, (key, value) => {
       const fnc = allOr ? "orWhere" : "where"
@@ -55,7 +56,7 @@ function addFilters(query, filters) {
   return query
 }
 
-function buildCreate(knex, json, opts) {
+function buildCreate(knex: Knex, json: QueryJson, opts: QueryOptions) {
   const { endpoint, body } = json
   let query = knex(endpoint.entityId)
   // mysql can't use returning
@@ -66,9 +67,9 @@ function buildCreate(knex, json, opts) {
   }
 }
 
-function buildRead(knex, json, limit) {
+function buildRead(knex: Knex, json: QueryJson, limit: number) {
   let { endpoint, resource, filters, sort, paginate } = json
-  let query = knex(endpoint.entityId)
+  let query: Knex.QueryBuilder = knex(endpoint.entityId)
   // select all if not specified
   if (!resource) {
     resource = { fields: [] }
@@ -90,6 +91,7 @@ function buildRead(knex, json, limit) {
   }
   // handle pagination
   if (paginate && paginate.page && paginate.limit) {
+    // @ts-ignore
     const page = paginate.page <= 1 ? 0 : paginate.page - 1
     const offset = page * paginate.limit
     query = query.offset(offset).limit(paginate.limit)
@@ -101,7 +103,7 @@ function buildRead(knex, json, limit) {
   return query
 }
 
-function buildUpdate(knex, json, opts) {
+function buildUpdate(knex: Knex, json: QueryJson, opts: QueryOptions) {
   const { endpoint, body, filters } = json
   let query = knex(endpoint.entityId)
   query = addFilters(query, filters)
@@ -113,7 +115,7 @@ function buildUpdate(knex, json, opts) {
   }
 }
 
-function buildDelete(knex, json, opts) {
+function buildDelete(knex: Knex, json: QueryJson, opts: QueryOptions) {
   const { endpoint, filters } = json
   let query = knex(endpoint.entityId)
   query = addFilters(query, filters)
@@ -126,20 +128,19 @@ function buildDelete(knex, json, opts) {
 }
 
 class SqlQueryBuilder {
+  private readonly client: any
+  private readonly limit: number
   // pass through client to get flavour of SQL
-  constructor(client, limit = BASE_LIMIT) {
-    this._client = client
-    this._limit = limit
+  constructor(client: any, limit: number = BASE_LIMIT) {
+    this.client = client
+    this.limit = limit
   }
 
   /**
    * @param json the input JSON structure from which an SQL query will be built.
    * @return {string} the operation that was found in the JSON.
    */
-  _operation(json) {
-    if (!json || !json.endpoint) {
-      return ""
-    }
+  _operation(json: QueryJson): Operation {
     return json.endpoint.operation
   }
 
@@ -149,21 +150,21 @@ class SqlQueryBuilder {
    * which for the sake of mySQL stops adding the returning statement to inserts, updates and deletes.
    * @return {{ sql: string, bindings: object }} the query ready to be passed to the driver.
    */
-  _query(json, opts = {}) {
-    const knex = require("knex")({ client: this._client })
+  _query(json: QueryJson, opts: QueryOptions = {}) {
+    const client = knex({ client: this.client })
     let query
     switch (this._operation(json)) {
-      case DataSourceOperation.CREATE:
-        query = buildCreate(knex, json, opts)
+      case Operation.CREATE:
+        query = buildCreate(client, json, opts)
         break
-      case DataSourceOperation.READ:
-        query = buildRead(knex, json, this._limit, opts)
+      case Operation.READ:
+        query = buildRead(client, json, this.limit)
         break
-      case DataSourceOperation.UPDATE:
-        query = buildUpdate(knex, json, opts)
+      case Operation.UPDATE:
+        query = buildUpdate(client, json, opts)
         break
-      case DataSourceOperation.DELETE:
-        query = buildDelete(knex, json, opts)
+      case Operation.DELETE:
+        query = buildDelete(client, json, opts)
         break
       default:
         throw `Operation type is not supported by SQL query builder`
